fix: use float64 for TorBox availability field
TorBox API returns availability as a float (e.g. 1.869) but the struct
had it typed as int, causing JSON unmarshal errors on every sync cycle.

Closes #5

+++ pkg/torbox/types.go
@@ -36,7 +36,7 @@ type Torrent struct {
 	DownloadFinished bool          `json:"download_finished"`
 	Files            []TorrentFile `json:"files"`
 	InactiveCheck    int           `json:"inactive_check"`
-	Availability     int           `json:"availability"`
+	Availability     float64       `json:"availability"`
 }
 
 // GetAddedAt parses the created_at timestamp.
